requirements.txt
django-annoying
django_enumfield>=2.0.0
djangorestframework
jsonschema
drf-extensions
toolz
psycopg2-binary

# for tests
coverage
Faker
responses
djangorestframework-simplejwt
hashids
freezegun
twine
beautifultable
pyfcm==2.0.5
